+++ packages/vscode-extension/.vscode-test.mjs
@@ -1,0 +1,11 @@
+import { defineConfig } from '@vscode/test-electron'
+
+export default defineConfig({
+  files: 'src/__tests__/integration/**/*.int.test.ts',
+  version: 'stable',
+  workspaceFolder: './',
+  mocha: {
+    ui: 'bdd',
+    timeout: 20000,
+  },
+})